Change deepcopy of dicts to dict.copy
Deep copying of Model objects was failing with Python 3.5, using the RefResolver'd schemas. Changing this to dict.copy seems to resolve this.

* Change copy.deepcopy(d) to dict.copy(d) in Model initialization

Not detected any adverse affects (all tests pass using Python 3), however if there was a specific reason that deepcopy() was used I've not figued out what that is.