Undo changes to the `warlock.model_factory` API
The changes introduced by commit 64a771d151a5defb9bca82042f ("Allow
resolver to be set on the Model") added a new keyword argument to the
`warlock.model_factory` function. Due to the way Python handles
positional arguments, the change alters the way that existing code is
interpreted.

Before the change,

    warlock.model_factory(schema_arg, snd_pos_arg)

would execute with

    kwargs == {
        'base_class': snd_pos_arg,
        'name': None,
    }

but after the change it's interpreted as

    kwargs == {
        'resolver': snd_pos_arg,
        'base_class'=model.Model,
        'name': None,
    }

This commit re-arranges the arguments to restore the previous behaviour
in an effort to not break existing code.