trivial: Update poetry.lock ordering
Seems poetry has changed the ordering of the poetry.lock file at some
point in the past (using poetry 1.1.13 from the repos on Fedora 36).
Fortunately, unlike in the past [1], it now versions this file so future
changes should be less likely (see the '[metadata] lock-version' value
added here). It's debatable that we should even be committing this for a
library like warlock [2], but as long as this is here we have to live
with it. Commit an "update" thus correcting this ordering. This will
allow us to see the actual diff in a future change.

[1] https://stackoverflow.com/q/65625717
[2] https://stackoverflow.com/q/61037557

Signed-off-by: Stephen Finucane <stephen@that.guru>