Use poetry-core as the build backend
Use poetry-core rather than poetry as the build backend.  This
is the recommended upstream solution, as it produces the same end result
while not requiring the tools to install the complete set of poetry
dependencies while building the wheel, effectively making the builds
much faster.

See "Why is this required?" at https://pypi.org/project/poetry-core/